+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [5.0.7] - Unreleased
+## [5.1.0] - Unreleased
 ### Fixed
  * Fixed message shown when configuration file (JSON or YAML) is not correct
  * Fixed exception handling in `devo/sender/data.py`